Refactor(app): 구조변경
