--- src/app/aside/index.ts
@@ -1,1 +1,0 @@
-export { Aside } from "./ui/Aside";
