--- src/app/header/index.ts
@@ -1,1 +1,0 @@
-export { Header } from "./ui/Header";
