--- src/app/header/types.ts
@@ -1,3 +1,0 @@
-export type LogoutModalProps = {
-    cancell: () => void;
-}
