+++ src/app/layout/ui/aside/ui/teamSpaceChannel/model/useOutsideChannel.ts
